requirements.txt
earwax
requests
attrs
